fixed: joining to group

+++ src/TeamManager/TeamManager/Controllers/UserGroupController.cs
@@ -1,42 +1,63 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TeamManager.Core.Entities;
 using TeamManager.Repository.Common;
-using System;
-using System.IO;
-using System.Threading.Tasks;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Identity;
 
 namespace TeamManager.Controllers
 {
-    [Authorize]
     public class UserGroupController : Controller
     {
         private readonly IRepository<UserGroup, Guid> _userGroupRepository;
-        private readonly UserManager<User> _userManager;
         private readonly IWebHostEnvironment _environment;
+        private readonly UserManager<User> _userManager;
 
         public UserGroupController(
             IRepository<UserGroup, Guid> userGroupRepository,
-            UserManager<User> userManager,
-            IWebHostEnvironment environment)
+            IWebHostEnvironment environment,
+            UserManager<User> userManager)
         {
             _userGroupRepository = userGroupRepository;
-            _userManager = userManager;
             _environment = environment;
+            _userManager = userManager;
+        }
+
+        private string GetCurrentUserId()
+        {
+            return User.FindFirstValue(ClaimTypes.NameIdentifier);
         }
 
         // GET: UserGroups
         public async Task<IActionResult> Index()
         {
-            var userGroups = await _userGroupRepository.GetAllAsync();
+            var currentUserId = GetCurrentUserId();
+            var isAdmin = User.IsInRole("Admin");
+            IEnumerable<UserGroup> userGroups;
+
+            if (isAdmin)
+            {
+                userGroups = await _userGroupRepository.GetAllAsync();
+            }
+            else
+            {
+                userGroups = (await _userGroupRepository.GetAllAsync())
+                              .Where(ug => ug.Users.Any(m => m.Id == Guid.Parse(currentUserId)));
+            }
+
             return View(userGroups);
         }
 
         // GET: UserGroups/Create
         public IActionResult Create()
         {
-            return View();
+            return View(new UserGroup());
         }
 
         // POST: UserGroups/Create
@@ -46,16 +67,8 @@ namespace TeamManager.Controllers
         {
             if (ModelState.IsValid)
             {
-                // Get the current logged-in user
-                var currentUser = await _userManager.GetUserAsync(User);
-                if (currentUser != null)
-                {
-                    userGroup.Users.Add(currentUser);
-                }
-
                 if (imageFile != null && imageFile.Length > 0)
                 {
-                    // Save uploaded image to wwwroot/img directory
                     var uploadsDir = Path.Combine(_environment.WebRootPath, "img");
                     var uniqueFileName = Guid.NewGuid().ToString() + "_" + imageFile.FileName;
                     var filePath = Path.Combine(uploadsDir, uniqueFileName);
@@ -66,9 +79,11 @@ namespace TeamManager.Controllers
                     userGroup.MainImage = "img/" + uniqueFileName;
                 }
 
+                userGroup.Id = Guid.NewGuid();
                 await _userGroupRepository.CreateAsync(userGroup);
                 return RedirectToAction(nameof(Index));
             }
+
             return View(userGroup);
         }
 
@@ -76,10 +91,11 @@ namespace TeamManager.Controllers
         public async Task<IActionResult> Edit(Guid id)
         {
             var userGroup = await _userGroupRepository.GetAsync(id);
-            if (userGroup == null)
+            if (userGroup == null || (!User.IsInRole("Admin") && !userGroup.Users.Any(m => m.Id == Guid.Parse(GetCurrentUserId()))))
             {
                 return NotFound();
             }
+
             return View(userGroup);
         }
 
@@ -95,40 +111,33 @@ namespace TeamManager.Controllers
 
             if (ModelState.IsValid)
             {
-                try
+                var currentUserId = GetCurrentUserId();
+                var userGroupToUpdate = await _userGroupRepository.GetAsync(id);
+                if (userGroupToUpdate == null || (!User.IsInRole("Admin") && !userGroupToUpdate.Users.Any(m => m.Id == Guid.Parse(currentUserId))))
                 {
-                    if (imageFile != null && imageFile.Length > 0)
-                    {
-                        // Save the new image file
-                        var uploadsDir = Path.Combine(_environment.WebRootPath, "uploads");
-                        if (!Directory.Exists(uploadsDir))
-                        {
-                            Directory.CreateDirectory(uploadsDir);
-                        }
-                        var uniqueFileName = Guid.NewGuid().ToString() + "_" + Path.GetFileName(imageFile.FileName);
-                        var filePath = Path.Combine(uploadsDir, uniqueFileName);
-                        using (var fileStream = new FileStream(filePath, FileMode.Create))
-                        {
-                            await imageFile.CopyToAsync(fileStream);
-                        }
-                        userGroup.MainImage = "/uploads/" + uniqueFileName;
-                    }
+                    return NotFound();
+                }
 
-                    await _userGroupRepository.UpdateAsync(userGroup);
-                }
-                catch (Exception)
+                userGroupToUpdate.Name = userGroup.Name;
+                userGroupToUpdate.Description = userGroup.Description;
+
+                if (imageFile != null && imageFile.Length > 0)
                 {
-                    if (!UserGroupExists(userGroup.Id))
+                    var uploadsDir = Path.Combine(_environment.WebRootPath, "img");
+                    var uniqueFileName = Guid.NewGuid().ToString() + "_" + imageFile.FileName;
+                    var filePath = Path.Combine(uploadsDir, uniqueFileName);
+                    using (var fileStream = new FileStream(filePath, FileMode.Create))
                     {
-                        return NotFound();
+                        await imageFile.CopyToAsync(fileStream);
                     }
-                    else
-                    {
-                        throw;
-                    }
+                    userGroupToUpdate.MainImage = "img/" + uniqueFileName;
                 }
+
+                await _userGroupRepository.UpdateAsync(userGroupToUpdate);
+
                 return RedirectToAction(nameof(Index));
             }
+
             return View(userGroup);
         }
 
@@ -136,7 +145,7 @@ namespace TeamManager.Controllers
         public async Task<IActionResult> Delete(Guid id)
         {
             var userGroup = await _userGroupRepository.GetAsync(id);
-            if (userGroup == null)
+            if (userGroup == null || (!User.IsInRole("Admin") && !userGroup.Users.Any(m => m.Id == Guid.Parse(GetCurrentUserId()))))
             {
                 return NotFound();
             }
@@ -148,13 +157,37 @@ namespace TeamManager.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+            var userGroup = await _userGroupRepository.GetAsync(id);
+            if (userGroup == null || (!User.IsInRole("Admin") && !userGroup.Users.Any(m => m.Id == Guid.Parse(GetCurrentUserId()))))
+            {
+                return NotFound();
+            }
+
             await _userGroupRepository.DeleteAsync(id);
             return RedirectToAction(nameof(Index));
         }
 
-        private bool UserGroupExists(Guid id)
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> LeaveGroup(Guid id)
         {
-            return _userGroupRepository.GetAsync(id) != null;
+            var userGroup = await _userGroupRepository.GetAsync(id);
+            if (userGroup == null)
+            {
+                return NotFound();
+            }
+
+            var currentUserId = Guid.Parse(GetCurrentUserId());
+            var currentUser = userGroup.Users.FirstOrDefault(u => u.Id == currentUserId);
+            if (currentUser == null)
+            {
+                return Forbid();
+            }
+
+            userGroup.Users.Remove(currentUser);
+            await _userGroupRepository.UpdateAsync(userGroup);
+
+            return RedirectToAction(nameof(Index));
         }
     }
 }
